Remove obsolete compatibility experiments
Amp-Thread-ID: https://ampcode.com/threads/T-019fefd8-cfd2-70a9-8c37-2a82c2b5bc13
