--- .buildkite/migration-poc-continuation.yml
@@ -1,14 +1,0 @@
-agents:
-  queue: "hosted"
-
-steps:
-  - label: ":white_check_mark: Migration POC suite settled"
-    key: "migration-poc-suite-terminal"
-    depends_on:
-      - step: "gha-poc-basic"
-        allow_failure: true
-      - step: "gha-poc-artifact-consumer"
-        allow_failure: true
-      - step: "gha-poc-advanced"
-        allow_failure: true
-    command: "echo 'Basic, artifact, and advanced migration POCs settled.'"
