--- .buildkite/migration-poc.yml
@@ -1,47 +1,0 @@
-agents:
-  queue: "hosted"
-
-steps:
-  - label: ":test_tube: Basic migration POC importer"
-    key: "migration-poc-basic-importer"
-    env:
-      BUILDKITE_GHA_TARGET_QUEUE: "hosted"
-    timeout_in_minutes: 15
-    retry:
-      automatic: false
-    plugins:
-      - mise#a5845c5082d3a4fe36dd77ae74973dfc86fc91a2:
-          version: "2026.5.12"
-    command: "scripts/migration-poc-import .github/workflows/example-basic.yml"
-
-  - label: ":package: Artifact migration POC importer"
-    key: "migration-poc-artifact-importer"
-    env:
-      BUILDKITE_GHA_TARGET_QUEUE: "hosted"
-    timeout_in_minutes: 15
-    retry:
-      automatic: false
-    plugins:
-      - mise#a5845c5082d3a4fe36dd77ae74973dfc86fc91a2:
-          version: "2026.5.12"
-    command: "scripts/migration-poc-import .github/workflows/example-artifacts.yml"
-
-  - label: ":rocket: Advanced migration POC importer"
-    key: "migration-poc-advanced-importer"
-    env:
-      BUILDKITE_GHA_TARGET_QUEUE: "hosted"
-    timeout_in_minutes: 15
-    retry:
-      automatic: false
-    plugins:
-      - mise#a5845c5082d3a4fe36dd77ae74973dfc86fc91a2:
-          version: "2026.5.12"
-    command: "scripts/migration-poc-import .github/workflows/example-advanced.yml"
-
-  - label: ":pipeline: Migration POC continuation loader"
-    key: "migration-poc-continuation-loader"
-    depends_on:
-      - "migration-poc-basic-importer"
-      - "migration-poc-artifact-importer"
-      - "migration-poc-advanced-importer"
-    command: "buildkite-agent pipeline upload .buildkite/migration-poc-continuation.yml"
