--- .buildkite/shell-oracle.yml
@@ -1,82 +1,0 @@
-agents:
-  queue: "hosted"
-
-steps:
-  - label: ":git: Materialize exact shell fixture"
-    key: shell-oracle-prepare
-    timeout_in_minutes: 10
-    command: |
-      set -euo pipefail
-      scripts/verify-source-checkout "$$ORACLE_SOURCE_COMMIT"
-      fixture_commit=$$(go run ./internal/harness/cmd/shell-oracle materialize --source testdata/smoke)
-      buildkite-agent meta-data set shell-oracle-fixture-commit "$$fixture_commit"
-      printf 'Materialized shell fixture commit: %s\n' "$$fixture_commit"
-    plugins:
-      - mise#a5845c5082d3a4fe36dd77ae74973dfc86fc91a2:
-          version: "2026.5.12"
-
-  - label: ":pipeline: Produce bounded shell output"
-    key: shell-oracle-producer
-    depends_on: shell-oracle-prepare
-    timeout_in_minutes: 5
-    command: |
-      set -euo pipefail
-      scripts/verify-source-checkout "$$ORACLE_SOURCE_COMMIT"
-      buildkite-agent meta-data set shell-oracle-result smoke-shell
-
-  - label: ":mag: Capture shell observation one"
-    key: shell-oracle-consumer-one
-    depends_on: shell-oracle-producer
-    timeout_in_minutes: 5
-    command: |
-      set -euo pipefail
-      scripts/verify-source-checkout "$$ORACLE_SOURCE_COMMIT"
-      result=$$(buildkite-agent meta-data get shell-oracle-result)
-      test "$$result" = "smoke-shell"
-      mkdir -p shell-oracle-observations
-      printf 'SMOKE_OBSERVATION={"result":"%s","variant":"one"}\n' "$$result" \
-        | tee shell-oracle-observations/observation-one.log
-      buildkite-agent artifact upload "shell-oracle-observations/observation-one.log"
-
-  - label: ":mag: Capture shell observation two"
-    key: shell-oracle-consumer-two
-    depends_on: shell-oracle-producer
-    timeout_in_minutes: 5
-    command: |
-      set -euo pipefail
-      scripts/verify-source-checkout "$$ORACLE_SOURCE_COMMIT"
-      result=$$(buildkite-agent meta-data get shell-oracle-result)
-      test "$$result" = "smoke-shell"
-      mkdir -p shell-oracle-observations
-      printf 'SMOKE_OBSERVATION={"result":"%s","variant":"two"}\n' "$$result" \
-        | tee shell-oracle-observations/observation-two.log
-      buildkite-agent artifact upload "shell-oracle-observations/observation-two.log"
-
-  - label: ":white_check_mark: Normalize and compare shell observations"
-    key: shell-oracle-compare
-    depends_on:
-      - shell-oracle-consumer-one
-      - shell-oracle-consumer-two
-    timeout_in_minutes: 10
-    command: |
-      set -euo pipefail
-      scripts/verify-source-checkout "$$ORACLE_SOURCE_COMMIT"
-      mkdir -p shell-oracle-observations
-      buildkite-agent artifact download \
-        "shell-oracle-observations/observation-one.log" . \
-        --step shell-oracle-consumer-one
-      buildkite-agent artifact download \
-        "shell-oracle-observations/observation-two.log" . \
-        --step shell-oracle-consumer-two
-      fixture_commit=$$(buildkite-agent meta-data get shell-oracle-fixture-commit)
-      cat \
-        shell-oracle-observations/observation-one.log \
-        shell-oracle-observations/observation-two.log \
-        | go run ./internal/harness/cmd/shell-oracle compare \
-            --source testdata/smoke \
-            --commit "$$fixture_commit" \
-            --provider buildkite \
-        | tee shell-oracle-normalized.json
-    plugins:
-      - mise#a5845c5082d3a4fe36dd77ae74973dfc86fc91a2:
-          version: "2026.5.12"
