--- .buildkite/transport-probe/fixtures/runtime-binding.claims.json
@@ -1,1 +1,0 @@
-{"build":{"build_id":"33333333-3333-4333-8333-333333333333","organization_id":"11111111-1111-4111-8111-111111111111","pipeline_id":"22222222-2222-4222-8222-222222222222"},"capability_ceiling":["network","secrets"],"event":{"attestation":"buildkite-webhook","commit":"0123456789abcdef0123456789abcdef01234567","digest":"sha256:b8e1f80bd70ae0784c7855a451731b745fddb67749d23f637be9082b75e9575b","name":"push","provider":"github","ref":"refs/heads/main","repository":"https://github.com/acme/widgets","trust":"trusted"},"exp":1784685600,"iat":1784682000,"iss":"buildkite-gha-plan-envelope","jti":"66666666-6666-4666-8666-666666666666","plan_digest":"sha256:64879f7d6b960a01909762d911a32d4582c20010c5641ee90278b644a9e3b525","queue":"gha-runtime","step_key":"gha-producer"}
