--- .buildkite/transport-probe/pipeline.yml
@@ -1,18 +1,0 @@
-agents:
-  queue: "hosted"
-
-steps:
-  - label: ":pipeline: Transport probe importer"
-    key: "transport-probe-importer"
-    command: ".buildkite/transport-probe/probe.sh bootstrap"
-    env:
-      TRANSPORT_PROBE_SIGN_JSON: "scripts/transport-probe-sign-json"
-      TRANSPORT_PROBE_VERIFY_JSON: "scripts/transport-probe-verify-json"
-    plugins:
-      - mise#a5845c5082d3a4fe36dd77ae74973dfc86fc91a2:
-          version: "2026.5.12"
-
-  - label: ":white_check_mark: Native dependency-extension check"
-    key: "transport-probe-native-after-import"
-    depends_on: "transport-probe-importer"
-    command: ".buildkite/transport-probe/probe.sh native"
