+++ internal/transport/model.go
@@ -10,6 +10,7 @@ import (
 	"encoding/json"
 	"errors"
 	"fmt"
+	"io"
 	"regexp"
 	"sort"
 	"strings"
@@ -35,7 +36,6 @@ var (
 	keyPattern                = regexp.MustCompile(`^[A-Za-z0-9_-]{1,255}$`)
 	logicalJobIDPattern       = regexp.MustCompile(`^[A-Za-z0-9_-]+(?:\.[A-Za-z0-9_-]+)*$`)
 	uuidPattern               = regexp.MustCompile(`^[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$`)
-	commitPattern             = regexp.MustCompile(`^[0-9a-f]{40}$`)
 	resultArtifactIDPattern   = regexp.MustCompile(`^[1-9][0-9]{0,19}$`)
 	resultArtifactPathPattern = regexp.MustCompile(`^buildkite-gha/v1/artifacts/[0-9a-f]{64}\.zip$`)
 )
@@ -46,115 +46,14 @@ func Digest(contents []byte) string {
 	return "sha256:" + hex.EncodeToString(sum[:])
 }
 
-// PlanArtifact is an immutable canonical plan and its compiler attribution.
-type PlanArtifact struct {
-	StepKey  string
-	Digest   string
-	Contents []byte
-	Binding  []byte
-}
-
-func (a PlanArtifact) validate() error {
-	if !keyPattern.MatchString(a.StepKey) {
-		return fmt.Errorf("invalid plan step key %q", a.StepKey)
-	}
-	if !digestPattern.MatchString(a.Digest) || Digest(a.Contents) != a.Digest {
-		return errors.New("plan digest does not match contents")
-	}
-	if len(a.Binding) == 0 {
-		return errors.New("signed plan binding is required")
+func requireJSONEOF(decoder *json.Decoder) error {
+	var extra any
+	if err := decoder.Decode(&extra); err == nil {
+		return errors.New("multiple JSON values")
+	} else if !errors.Is(err, io.EOF) {
+		return errors.New("invalid trailing JSON data")
 	}
 	return nil
-}
-
-// Path is collision-resistant across plans and stable across retries.
-func (a PlanArtifact) Path() string {
-	return fmt.Sprintf("buildkite-gha/v1/plans/%s/%s/plan.json", a.StepKey, strings.TrimPrefix(a.Digest, "sha256:"))
-}
-
-// BindingPath is the signed transport integrity artifact beside the inert plan.
-func (a PlanArtifact) BindingPath() string {
-	return fmt.Sprintf("buildkite-gha/v1/plans/%s/%s/binding.jws", a.StepKey, strings.TrimPrefix(a.Digest, "sha256:"))
-}
-
-// Dependency separates the strict compiler edge from logical GHA needs.
-type Dependency struct {
-	StepKey string
-}
-
-// Job is the validated subset needed to emit one generated Buildkite job.
-type Job struct {
-	Key          string
-	Label        string
-	Command      string
-	Queue        string
-	Plan         PlanArtifact
-	Dependencies []Dependency
-}
-
-// EmitTwoJobPipeline emits the deterministic transport-probe producer/consumer pipeline.
-// The compiler dependency is injected as strict; all supplied logical needs
-// are forced to settle even when their producer failed.
-func EmitTwoJobPipeline(compilerStep string, jobs []Job) ([]byte, error) {
-	if !keyPattern.MatchString(compilerStep) {
-		return nil, fmt.Errorf("invalid compiler step key %q", compilerStep)
-	}
-	if len(jobs) != 2 {
-		return nil, fmt.Errorf("transport probe requires exactly two generated jobs, got %d", len(jobs))
-	}
-	seen := map[string]bool{}
-	var out bytes.Buffer
-	out.WriteString("steps:\n")
-	for _, job := range jobs {
-		if err := validateJob(job, compilerStep, seen); err != nil {
-			return nil, err
-		}
-		seen[job.Key] = true
-		_, _ = fmt.Fprintf(&out, "  - label: %s\n", yamlString(job.Label))
-		_, _ = fmt.Fprintf(&out, "    key: %s\n", yamlString(job.Key))
-		_, _ = fmt.Fprintf(&out, "    command: %s\n", yamlString(job.Command))
-		_, _ = fmt.Fprintf(&out, "    agents:\n      queue: %s\n", yamlString(job.Queue))
-		out.WriteString("    checkout:\n      skip: true\n")
-		out.WriteString("    env:\n")
-		_, _ = fmt.Fprintf(&out, "      BUILDKITE_GHA_PLAN_DIGEST: %s\n", yamlString(job.Plan.Digest))
-		_, _ = fmt.Fprintf(&out, "      BUILDKITE_GHA_PLAN_PRODUCER: %s\n", yamlString(compilerStep))
-		out.WriteString("    depends_on:\n")
-		_, _ = fmt.Fprintf(&out, "      - step: %s\n        allow_failure: false\n", yamlString(compilerStep))
-		dependencies := append([]Dependency(nil), job.Dependencies...)
-		sort.Slice(dependencies, func(i, j int) bool { return dependencies[i].StepKey < dependencies[j].StepKey })
-		for _, dependency := range dependencies {
-			_, _ = fmt.Fprintf(&out, "      - step: %s\n        allow_failure: true\n", yamlString(dependency.StepKey))
-		}
-	}
-	return out.Bytes(), nil
-}
-
-func validateJob(job Job, compilerStep string, seen map[string]bool) error {
-	if !keyPattern.MatchString(job.Key) || job.Key == compilerStep || seen[job.Key] {
-		return fmt.Errorf("invalid or duplicate generated step key %q", job.Key)
-	}
-	if job.Label == "" || job.Command == "" || !keyPattern.MatchString(job.Queue) {
-		return fmt.Errorf("job %q requires label, command, and valid queue", job.Key)
-	}
-	if err := job.Plan.validate(); err != nil {
-		return fmt.Errorf("job %q: %w", job.Key, err)
-	}
-	if job.Plan.StepKey != job.Key {
-		return fmt.Errorf("job %q plan is bound to %q", job.Key, job.Plan.StepKey)
-	}
-	dependencies := map[string]bool{}
-	for _, dependency := range job.Dependencies {
-		if !seen[dependency.StepKey] || dependencies[dependency.StepKey] {
-			return fmt.Errorf("job %q has unknown, forward, or duplicate need %q", job.Key, dependency.StepKey)
-		}
-		dependencies[dependency.StepKey] = true
-	}
-	return nil
-}
-
-func yamlString(value string) string {
-	encoded, _ := json.Marshal(value)
-	return string(encoded)
 }
 
 // Output is a bounded, non-secret logical job output.
