+++ internal/transport/model_test.go
@@ -18,72 +18,6 @@ const (
 	testJobID2  = "55555555-5555-4555-8555-555555555555"
 )
 
-func TestEmitTwoJobPipelineIsDeterministicAndKeepsDependencyPoliciesSeparate(t *testing.T) {
-	producer := testPlan("gha-producer", `{"job":"producer"}`)
-	consumer := testPlan("gha-consumer", `{"job":"consumer"}`)
-	jobs := []Job{
-		{Key: "gha-producer", Label: "Producer", Command: "probe producer", Queue: "gha-runtime", Plan: producer},
-		{Key: "gha-consumer", Label: "Consumer", Command: "probe consumer", Queue: "gha-runtime", Plan: consumer, Dependencies: []Dependency{{StepKey: "gha-producer"}}},
-	}
-	first, err := EmitTwoJobPipeline("gha-importer", jobs)
-	if err != nil {
-		t.Fatal(err)
-	}
-	second, err := EmitTwoJobPipeline("gha-importer", jobs)
-	if err != nil {
-		t.Fatal(err)
-	}
-	if !bytes.Equal(first, second) {
-		t.Fatal("pipeline emission was not byte-identical")
-	}
-	want := `steps:
-  - label: "Producer"
-    key: "gha-producer"
-    command: "probe producer"
-    agents:
-      queue: "gha-runtime"
-    checkout:
-      skip: true
-    env:
-      BUILDKITE_GHA_PLAN_DIGEST: "` + producer.Digest + `"
-      BUILDKITE_GHA_PLAN_PRODUCER: "gha-importer"
-    depends_on:
-      - step: "gha-importer"
-        allow_failure: false
-  - label: "Consumer"
-    key: "gha-consumer"
-    command: "probe consumer"
-    agents:
-      queue: "gha-runtime"
-    checkout:
-      skip: true
-    env:
-      BUILDKITE_GHA_PLAN_DIGEST: "` + consumer.Digest + `"
-      BUILDKITE_GHA_PLAN_PRODUCER: "gha-importer"
-    depends_on:
-      - step: "gha-importer"
-        allow_failure: false
-      - step: "gha-producer"
-        allow_failure: true
-`
-	if string(first) != want {
-		t.Fatalf("pipeline changed\nwant:\n%s\ngot:\n%s", want, first)
-	}
-}
-
-func TestEmitTwoJobPipelineRejectsForwardNeed(t *testing.T) {
-	producer := testPlan("gha-producer", `{}`)
-	consumer := testPlan("gha-consumer", `{}`)
-	jobs := []Job{
-		{Key: "gha-producer", Label: "Producer", Command: "true", Queue: "runtime", Plan: producer, Dependencies: []Dependency{{StepKey: "gha-consumer"}}},
-		{Key: "gha-consumer", Label: "Consumer", Command: "true", Queue: "runtime", Plan: consumer},
-	}
-	_, err := EmitTwoJobPipeline("gha-importer", jobs)
-	if err == nil || !strings.Contains(err.Error(), "unknown, forward, or duplicate need") {
-		t.Fatalf("error = %v", err)
-	}
-}
-
 func TestResultManifestIsCanonicalAndProducerBound(t *testing.T) {
 	producer := Producer{BuildID: testBuildID, JobID: testJobID, StepKey: "gha-producer"}
 	manifest := ResultManifest{
@@ -380,104 +314,6 @@ func TestUploadArtifactsFailsBeforePipeline(t *testing.T) {
 	}
 }
 
-func TestUploadOrdersArtifactsMarkersAndPipelineAndFailsClosed(t *testing.T) {
-	key := testKey(t)
-	producer := testPlan("gha-producer", `{"job":"producer"}`)
-	consumer := testPlan("gha-consumer", `{"job":"consumer"}`)
-	pipeline := []byte("steps: []\n")
-	intent := UploadIntent{
-		BuildID: testBuildID, ImporterKey: "gha-importer", PipelineDigest: Digest(pipeline),
-		Jobs: []UploadJob{{Key: producer.StepKey, PlanDigest: producer.Digest}, {Key: consumer.StepKey, PlanDigest: consumer.Digest}},
-	}
-	expected, err := SignMarker(key, intent, "expected")
-	if err != nil {
-		t.Fatal(err)
-	}
-	completed, err := SignMarker(key, intent, "completed")
-	if err != nil {
-		t.Fatal(err)
-	}
-	runner := &captureRunner{}
-	root := t.TempDir()
-	if err := os.MkdirAll(filepath.Dir(filepath.Join(root, filepath.FromSlash(consumer.Path()))), 0o700); err != nil {
-		t.Fatal(err)
-	}
-	if err := os.WriteFile(filepath.Join(root, filepath.FromSlash(consumer.Path())), []byte("stale plan"), 0o600); err != nil {
-		t.Fatal(err)
-	}
-	if err := Upload(context.Background(), Agent{Runner: runner}, root, []PlanArtifact{consumer, producer}, pipeline, expected, completed); err != nil {
-		t.Fatal(err)
-	}
-	wantKinds := []string{"artifact", "artifact", "artifact", "artifact", "meta-data", "pipeline", "meta-data"}
-	for i, kind := range wantKinds {
-		if runner.commands[i].args[0] != kind {
-			t.Fatalf("command %d = %#v, want %s", i, runner.commands[i], kind)
-		}
-	}
-	resolvedRoot, err := filepath.EvalSymlinks(root)
-	if err != nil {
-		t.Fatal(err)
-	}
-	path := func(relative string) string { return filepath.Join(resolvedRoot, filepath.FromSlash(relative)) }
-	if runner.commands[0].dir != resolvedRoot || runner.commands[1].dir != resolvedRoot || runner.commands[2].dir != resolvedRoot || runner.commands[3].dir != resolvedRoot || runner.commands[0].args[2] != consumer.Path() || runner.commands[1].args[2] != consumer.BindingPath() || runner.commands[2].args[2] != producer.Path() || runner.commands[3].args[2] != producer.BindingPath() {
-		t.Fatalf("plans and bindings not sorted before upload: %#v", runner.commands[:4])
-	}
-	for _, plan := range []PlanArtifact{producer, consumer} {
-		contents, err := os.ReadFile(path(plan.Path()))
-		if err != nil || !bytes.Equal(contents, plan.Contents) {
-			t.Fatalf("materialized plan %q = %q, %v", plan.StepKey, contents, err)
-		}
-		binding, err := os.ReadFile(path(plan.BindingPath()))
-		if err != nil || !bytes.Equal(binding, plan.Binding) {
-			t.Fatalf("materialized binding %q = %q, %v", plan.StepKey, binding, err)
-		}
-	}
-
-	failing := &captureRunner{failAt: 6}
-	err = Upload(context.Background(), Agent{Runner: failing}, t.TempDir(), []PlanArtifact{producer, consumer}, pipeline, expected, completed)
-	if err == nil || len(failing.commands) != 6 {
-		t.Fatalf("Upload() error = %v, commands = %d; completion marker must not be published", err, len(failing.commands))
-	}
-
-	missing := &captureRunner{}
-	err = Upload(context.Background(), Agent{Runner: missing}, t.TempDir(), []PlanArtifact{producer}, pipeline, expected, completed)
-	if err == nil || len(missing.commands) != 0 {
-		t.Fatalf("Upload() error = %v, commands = %d; missing signed plan set must fail before side effects", err, len(missing.commands))
-	}
-}
-
-func TestUploadRejectsTamperedMaterializedBindingBeforeAgentReadsIt(t *testing.T) {
-	key := testKey(t)
-	producer := testPlan("gha-producer", `{"job":"producer"}`)
-	consumer := testPlan("gha-consumer", `{"job":"consumer"}`)
-	pipeline := []byte("steps: []\n")
-	intent := UploadIntent{
-		BuildID: testBuildID, ImporterKey: "gha-importer", PipelineDigest: Digest(pipeline),
-		Jobs: []UploadJob{{Key: producer.StepKey, PlanDigest: producer.Digest}, {Key: consumer.StepKey, PlanDigest: consumer.Digest}},
-	}
-	expected, err := SignMarker(key, intent, "expected")
-	if err != nil {
-		t.Fatal(err)
-	}
-	completed, err := SignMarker(key, intent, "completed")
-	if err != nil {
-		t.Fatal(err)
-	}
-	root := t.TempDir()
-	tampered := filepath.Join(root, filepath.FromSlash(consumer.BindingPath()))
-	runner := &captureRunner{afterRun: func(command int) {
-		if command == 1 {
-			if err := os.WriteFile(tampered, []byte("tampered"), 0o600); err != nil {
-				t.Fatal(err)
-			}
-		}
-	}}
-	err = Upload(context.Background(), Agent{Runner: runner}, root, []PlanArtifact{producer, consumer}, pipeline, expected, completed)
-	if err == nil || !strings.Contains(err.Error(), "on-disk bytes differ") || len(runner.commands) != 1 {
-		t.Fatalf("Upload() error = %v, commands = %d", err, len(runner.commands))
-	}
-}
-
 func TestCommandRunnerBoundsOutputWhileReading(t *testing.T) {
 	stdout, _, err := (CommandRunner{}).RunBounded(context.Background(), "", "sh", []string{"-c", "printf 123456789"}, nil, 8)
 	if err == nil || !strings.Contains(err.Error(), "exceeds 8 bytes") || len(stdout) != 0 {
@@ -514,18 +350,4 @@ func TestGetMetadataBoundedClassifiesOnlyExpectedWebhookAbsence(t *testing.T) {
 			}
 		})
 	}
-}
-
-func testPlan(stepKey, contents string) PlanArtifact {
-	data := []byte(contents)
-	return PlanArtifact{StepKey: stepKey, Digest: Digest(data), Contents: data, Binding: []byte("signed-binding")}
-}
-
-func testKey(t *testing.T) ES256Key {
-	t.Helper()
-	key, err := NewTestES256Key("transport-probe-test")
-	if err != nil {
-		t.Fatal(err)
-	}
-	return key
 }
