+++ mise.toml
@@ -34,7 +34,7 @@ run = "golangci-lint run"
 
 [tasks."lint:shell"]
 description = "Lint shell scripts"
-run = "shellcheck -x .buildkite/transport-probe/probe.sh .buildkite/upload-examples.sh scripts/ci-buildkite-release scripts/compare-example scripts/migration-poc-import scripts/oss-corpus scripts/verify-source-checkout scripts/transport-probe-sign-json scripts/transport-probe-verify-json scripts/hosted-docker-probe scripts/container-runtime-probe scripts/verify-summary-annotation scripts/verify-workflow-annotations scripts/verify-upload-artifact scripts/verify-artifact-roundtrip scripts/release scripts/smoke-local testdata/actions/docker/entrypoint.sh testdata/container-runtime/.github/actions/docker/entrypoint.sh"
+run = "shellcheck -x .buildkite/upload-examples.sh scripts/ci-buildkite-release scripts/compare-example scripts/oss-corpus scripts/verify-source-checkout scripts/hosted-docker-probe scripts/container-runtime-probe scripts/verify-summary-annotation scripts/verify-workflow-annotations scripts/verify-upload-artifact scripts/verify-artifact-roundtrip scripts/release scripts/smoke-local testdata/actions/docker/entrypoint.sh testdata/container-runtime/.github/actions/docker/entrypoint.sh"
 
 [tasks.vet]
 description = "Run go vet"
@@ -52,10 +52,6 @@ run = "scripts/release"
 [tasks."release:check"]
 description = "Validate the release configuration"
 run = "goreleaser check"
-
-[tasks.plan-fixtures]
-description = "Validate signed plan-envelope fixtures"
-run = "python3 testdata/plans/validate.py"
 
 [tasks."smoke:local"]
 description = "Validate and deterministically compile the local smoke inventory"
@@ -75,4 +71,4 @@ run = "./scripts/oss-corpus --profile"
 
 [tasks.check]
 description = "Run formatting, build, tests, lint, vet, and fixture checks"
-run = "mise run --jobs 1 format ::: build ::: test ::: test:race ::: lint:go ::: lint:shell ::: vet ::: plan-fixtures ::: smoke:local ::: release:check"
+run = "mise run --jobs 1 format ::: build ::: test ::: test:race ::: lint:go ::: lint:shell ::: vet ::: smoke:local ::: release:check"
