--- schemas/plan-envelope-protected-header-v1.schema.json
@@ -1,13 +1,0 @@
-{
-  "$schema": "https://json-schema.org/draft/2020-12/schema",
-  "$id": "https://buildkite.com/schemas/buildkite-gha/plan-envelope-protected-header-v1.schema.json",
-  "title": "buildkite-gha plan-envelope JWS protected header v1",
-  "type": "object",
-  "additionalProperties": false,
-  "required": ["alg", "kid", "typ"],
-  "properties": {
-    "alg": {"const": "ES256"},
-    "kid": {"type": "string", "pattern": "^[A-Za-z0-9._:/-]+$", "minLength": 1, "maxLength": 512},
-    "typ": {"const": "buildkite-gha-plan-envelope+jws"}
-  }
-}
