--- scripts/migration-poc-import
@@ -1,56 +1,0 @@
-#!/usr/bin/env bash
-set -euo pipefail
-
-if (( $# != 1 )); then
-  echo 'usage: scripts/migration-poc-import <workflow>' >&2
-  exit 2
-fi
-
-workflow=$1
-case "$workflow" in
-  .github/workflows/example-basic.yml | \
-    .github/workflows/example-artifacts.yml | \
-    .github/workflows/example-advanced.yml) ;;
-  *)
-    echo "unsupported migration POC workflow: $workflow" >&2
-    exit 2
-    ;;
-esac
-
-commit=${POC_COMMIT:?POC_COMMIT is required}
-scripts/verify-source-checkout "$commit"
-test "${BUILDKITE_COMMIT:?BUILDKITE_COMMIT is required}" = "$commit"
-test -z "$(git status --porcelain --untracked-files=all)"
-
-distribution_root=$(mktemp -d "${TMPDIR:-/tmp}/buildkite-gha-migration-poc.XXXXXXXX")
-event_path=$(mktemp "${TMPDIR:-/tmp}/buildkite-gha-migration-event.XXXXXXXX.json")
-trap 'rm -rf -- "$distribution_root"; rm -f -- "$event_path"' EXIT
-
-branch=${BUILDKITE_BRANCH:?BUILDKITE_BRANCH is required}
-jq -n \
-  --arg ref "refs/heads/$branch" \
-  --arg sha "$commit" \
-  '{
-    provider: "github",
-    event: "push",
-    repository: {
-      owner: "buildkite",
-      name: "buildkite-gha",
-      clone_url: "https://github.com/buildkite/buildkite-gha.git",
-      default_branch: "main"
-    },
-    ref: $ref,
-    sha: $sha,
-    actor: "buildkite-gha-migration-poc",
-    payload: {ref: $ref}
-  }' > "$event_path"
-
-runtime_version="0.0.0-poc.$commit"
-CGO_ENABLED=0 mise exec -- go build -trimpath -buildvcs=false \
-  -ldflags "-X main.version=$runtime_version" \
-  -o "$distribution_root/buildkite-gha" ./cmd/buildkite-gha
-
-"$distribution_root/buildkite-gha" upload \
-  --event-path "$event_path" \
-  --runtime-queue hosted \
-  "$workflow"
