--- scripts/transport-probe-sign-json
@@ -1,4 +1,0 @@
-#!/usr/bin/env bash
-set -euo pipefail
-
-mise exec -- go run ./internal/harness/cmd/transport-probe-signing sign
