--- testdata/plans/cases.json
@@ -1,145 +1,0 @@
-[
-  {
-    "envelope": "envelopes/valid.json",
-    "expected": "OK",
-    "name": "valid",
-    "plan": "plans/valid.json",
-    "runtime": {
-      "build_id": "33333333-3333-4333-8333-333333333333",
-      "local_capability_ceiling": [
-        "network",
-        "secrets"
-      ],
-      "now": 1784682060,
-      "organization_id": "11111111-1111-4111-8111-111111111111",
-      "pipeline_id": "22222222-2222-4222-8222-222222222222",
-      "queue": "gha-trusted",
-      "step_key": "gha-ci-test"
-    }
-  },
-  {
-    "envelope": "envelopes/valid.json",
-    "expected": "E_PLAN_DIGEST",
-    "name": "tampered",
-    "plan": "plans/tampered.json",
-    "runtime": {
-      "build_id": "33333333-3333-4333-8333-333333333333",
-      "local_capability_ceiling": [
-        "network",
-        "secrets"
-      ],
-      "now": 1784682060,
-      "organization_id": "11111111-1111-4111-8111-111111111111",
-      "pipeline_id": "22222222-2222-4222-8222-222222222222",
-      "queue": "gha-trusted",
-      "step_key": "gha-ci-test"
-    }
-  },
-  {
-    "envelope": "envelopes/valid.json",
-    "expected": "E_BUILD_BINDING",
-    "name": "wrong-build",
-    "plan": "plans/valid.json",
-    "runtime": {
-      "build_id": "44444444-4444-4444-8444-444444444444",
-      "local_capability_ceiling": [
-        "network",
-        "secrets"
-      ],
-      "now": 1784682060,
-      "organization_id": "11111111-1111-4111-8111-111111111111",
-      "pipeline_id": "22222222-2222-4222-8222-222222222222",
-      "queue": "gha-trusted",
-      "step_key": "gha-ci-test"
-    }
-  },
-  {
-    "envelope": "envelopes/valid.json",
-    "expected": "E_STEP_BINDING",
-    "name": "wrong-job",
-    "plan": "plans/valid.json",
-    "runtime": {
-      "build_id": "33333333-3333-4333-8333-333333333333",
-      "local_capability_ceiling": [
-        "network",
-        "secrets"
-      ],
-      "now": 1784682060,
-      "organization_id": "11111111-1111-4111-8111-111111111111",
-      "pipeline_id": "22222222-2222-4222-8222-222222222222",
-      "queue": "gha-trusted",
-      "step_key": "gha-ci-deploy"
-    }
-  },
-  {
-    "envelope": "envelopes/valid.json",
-    "expected": "E_QUEUE_BINDING",
-    "name": "wrong-queue",
-    "plan": "plans/valid.json",
-    "runtime": {
-      "build_id": "33333333-3333-4333-8333-333333333333",
-      "local_capability_ceiling": [
-        "network",
-        "secrets"
-      ],
-      "now": 1784682060,
-      "organization_id": "11111111-1111-4111-8111-111111111111",
-      "pipeline_id": "22222222-2222-4222-8222-222222222222",
-      "queue": "gha-privileged",
-      "step_key": "gha-ci-test"
-    }
-  },
-  {
-    "envelope": "envelopes/valid.json",
-    "expected": "E_EXPIRED",
-    "name": "expired",
-    "plan": "plans/valid.json",
-    "runtime": {
-      "build_id": "33333333-3333-4333-8333-333333333333",
-      "local_capability_ceiling": [
-        "network",
-        "secrets"
-      ],
-      "now": 1784685600,
-      "organization_id": "11111111-1111-4111-8111-111111111111",
-      "pipeline_id": "22222222-2222-4222-8222-222222222222",
-      "queue": "gha-trusted",
-      "step_key": "gha-ci-test"
-    }
-  },
-  {
-    "envelope": "envelopes/untrusted-event.json",
-    "expected": "E_CAPABILITY_POLICY",
-    "name": "untrusted-event",
-    "plan": "plans/valid.json",
-    "runtime": {
-      "build_id": "33333333-3333-4333-8333-333333333333",
-      "local_capability_ceiling": [
-        "network"
-      ],
-      "now": 1784682060,
-      "organization_id": "11111111-1111-4111-8111-111111111111",
-      "pipeline_id": "22222222-2222-4222-8222-222222222222",
-      "queue": "gha-trusted",
-      "step_key": "gha-ci-test"
-    }
-  },
-  {
-    "envelope": "envelopes/untrusted-key.json",
-    "expected": "E_UNTRUSTED_KEY",
-    "name": "untrusted-key",
-    "plan": "plans/valid.json",
-    "runtime": {
-      "build_id": "33333333-3333-4333-8333-333333333333",
-      "local_capability_ceiling": [
-        "network",
-        "secrets"
-      ],
-      "now": 1784682060,
-      "organization_id": "11111111-1111-4111-8111-111111111111",
-      "pipeline_id": "22222222-2222-4222-8222-222222222222",
-      "queue": "gha-trusted",
-      "step_key": "gha-ci-test"
-    }
-  }
-]
