--- testdata/plans/envelopes/untrusted-event.json
@@ -1,44 +1,0 @@
-{
-  "claims": {
-    "authorization": {
-      "capability_ceiling": [
-        "network",
-        "secrets"
-      ]
-    },
-    "build": {
-      "build_id": "33333333-3333-4333-8333-333333333333",
-      "organization_id": "11111111-1111-4111-8111-111111111111",
-      "pipeline_id": "22222222-2222-4222-8222-222222222222"
-    },
-    "exp": 1784685600,
-    "iat": 1784682000,
-    "iss": "buildkite-gha-plan-envelope",
-    "jti": "66666666-6666-4666-8666-666666666666",
-    "plan": {
-      "compiler_distribution_digest": "sha256:9f0417f050cd0ceeed8a1264e3957f6de57f3be9c3125bbbfe2fdd281a599ce7",
-      "compiler_version": "0.1.0",
-      "digest": "sha256:d5115ad7e6e537ef83c01478d3c5d9b8045664d3461d5be64626f3f15c36e1b5",
-      "schema": "https://buildkite.com/schemas/buildkite-gha/job-plan-v1.schema.json"
-    },
-    "provenance": {
-      "attestation": "buildkite-webhook",
-      "commit": "0123456789abcdef0123456789abcdef01234567",
-      "event_id": "evt-123",
-      "event_name": "push",
-      "event_payload_digest": "sha256:8db4617656fc634dd11e00067893a6d5c5e002f71bf12f66b7b09d0642521531",
-      "provider": "github",
-      "ref": "refs/heads/main",
-      "repository": "https://github.com/acme/widgets",
-      "trust": "untrusted",
-      "workflow_digest": "sha256:74d276a4cc8c7b0571783dbb97c7013f35a4f204821fafab429b1d940a7e45d4",
-      "workflow_path": ".github/workflows/ci.yml"
-    },
-    "target": {
-      "queue": "gha-trusted",
-      "step_key": "gha-ci-test"
-    }
-  },
-  "protected": "eyJhbGciOiJFUzI1NiIsImtpZCI6InBsYW4tZml4dHVyZS0yMDI2LTA3LTIyIiwidHlwIjoiYnVpbGRraXRlLWdoYS1wbGFuLWVudmVsb3BlK2p3cyJ9",
-  "signature": "pw4R1S0x20CiETbf4YkTOrocHCyEj_2gpeSAnqpjDVrlyLTljg8q64pJB0EqKW_Ca69wSimNus1-vmxIRwnFkA"
-}
