--- testdata/plans/plans/valid.json
@@ -1,1 +1,0 @@
-{"compiler":{"distribution_digest":"sha256:9f0417f050cd0ceeed8a1264e3957f6de57f3be9c3125bbbfe2fdd281a599ce7","version":"0.1.0"},"event":{"name":"push","payload_digest":"sha256:8db4617656fc634dd11e00067893a6d5c5e002f71bf12f66b7b09d0642521531","provider":"github"},"required_capabilities":["network","secrets"],"schema":"https://buildkite.com/schemas/buildkite-gha/job-plan-v1.schema.json","steps":[{"command":"make test","id":"test","kind":"run"}],"target":{"queue":"gha-trusted","step_key":"gha-ci-test"},"workflow":{"digest":"sha256:74d276a4cc8c7b0571783dbb97c7013f35a4f204821fafab429b1d940a7e45d4","logical_job_id":"test","path":".github/workflows/ci.yml"}}
