--- testdata/plans/revoked-kids.json
@@ -1,1 +1,0 @@
-[]
