--- testdata/plans/trust-roots.jwks.json
@@ -1,16 +1,0 @@
-{
-  "keys": [
-    {
-      "alg": "ES256",
-      "crv": "P-256",
-      "key_ops": [
-        "verify"
-      ],
-      "kid": "plan-fixture-2026-07-22",
-      "kty": "EC",
-      "use": "sig",
-      "x": "_bHKBz9Cza76zTLGTzIfEzjIPKQq0j6-EIb2JsqySog",
-      "y": "mFiQc8RJZiwpbxwjKClgy_vNDIMslWhy6Dwjw-UJZBg"
-    }
-  ]
-}
