--- testdata/smoke/expected/concurrent-capture.json
@@ -1,16 +1,0 @@
-{
-  "observations": [
-    {
-      "identity": "concurrent",
-      "document": {
-        "cancel": "graceful",
-        "failure": "failure-at-wait",
-        "implicit": "implicit-wait-all",
-        "parallel": "parallel",
-        "queue_max": 10,
-        "targeted": "targeted-and-full"
-      }
-    }
-  ],
-  "lifecycle": []
-}
