--- testdata/smoke/expected/shell-capture.json
@@ -1,19 +1,0 @@
-{
-  "observations": [
-    {
-      "identity": "consumer[variant=one]",
-      "document": {
-        "result": "smoke-shell",
-        "variant": "one"
-      }
-    },
-    {
-      "identity": "consumer[variant=two]",
-      "document": {
-        "result": "smoke-shell",
-        "variant": "two"
-      }
-    }
-  ],
-  "lifecycle": []
-}
